Update UI and functionality in `Hero.tsx`, `Install.tsx`, and `CopyCommand`
In `Hero.tsx`, the `className` for a `h1` tag was updated to adjust the minimum height at different screen resolutions. The `letterWrapClass` was also updated to include a new animation class, "fade-in-1".

In `Install.tsx`, several updates were made. The gap between items in a flex container was increased. The `className` for a `p` tag and a `div` tag were updated to adjust the font weight and padding respectively. The `defaultValue` for a `Tabs` component was changed and the value and text for a `TabsTrigger` component were updated to change the tab labels.

In `CopyCommand`, the `className` for a `div` tag was updated to adjust the width and padding. The `className` for a `code` tag was updated to change the display property and add a truncate class. The `className` for a `DropdownMenuContent` was also updated to remove an extra space.