Add `Accordion` and `Input` components to UI
This update includes the addition of two new components to the UI: `Accordion` and `Input`.

The `Accordion` component is a versatile, expandable content area that can be used to display or hide information. It has been implemented using the `@radix-ui/react-accordion` package, and includes `AccordionItem`, `AccordionTrigger`, and `AccordionContent` subcomponents.

The `Input` component is a basic form input field. It accepts all standard HTML input attributes and can be styled using the `className` prop.

In addition to these components, the `@radix-ui/react-accordion` dependency has been added to the project's `package.json` file.